添加mouse事件

+++ pEngine/src/pEngine/event/MouseEvent.h
@@ -8,30 +8,30 @@ namespace pEngine
 	class DLL_API MouseMoveEvent : public Event
 	{
 	public:
-		MouseMoveEvent(float x, float y) :
+		MouseMoveEvent(double x, double y) :
 			m_x(x), m_y(y) {}
-		inline int GetX() const { return m_x; }
-		inline int GetY() const { return m_y; }
+		inline double GetX() const { return m_x; }
+		inline double GetY() const { return m_y; }
 
 		EVENT_CLASS_TYEP(MouseMoved);
 	private:
-		float m_x;
-		float m_y;
+		double m_x;
+		double m_y;
 	};
 
 	class DLL_API MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(float xOffset, float yOffset):
+		MouseScrolledEvent(double xOffset, double yOffset):
 			m_xOffset(xOffset), m_yOffset(yOffset) {}
 
-		inline float GetxOffset() const { return m_xOffset; }
-		inline float GetyOffset() const { return m_yOffset; }
+		inline double GetxOffset() const { return m_xOffset; }
+		inline double GetyOffset() const { return m_yOffset; }
 
 		EVENT_CLASS_TYEP(MouseScrolled);
 	private:
-		float m_xOffset;
-		float m_yOffset;
+		double m_xOffset;
+		double m_yOffset;
 	};
 
 	class DLL_API MouseButtonEvnet : public Event
